+++ docs/examples/controlpanels/post_content_rules.ts
@@ -3,7 +3,7 @@ import { Client } from '@robgietema/nick';
 const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
 const login = await cli.login({ username: 'admin', password: 'admin' });
 
-const { data } = await cli.createControlpanelType({
+const { data } = await cli.createControlpanelContentRule({
   token: login.data.token,
   data: {
     title: 'Content Rule 3',