+++ docs/examples/controlpanels/post_content_rules_condition.ts
@@ -1,0 +1,12 @@
+import { Client } from '@robgietema/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({ username: 'admin', password: 'admin' });
+
+const { data } = await cli.createControlpanelContentRuleCondition({
+  token: login.data.token,
+  data: {
+    file_extension: 'jpg',
+    type: 'file_extension',
+  },
+});