+++ src/routes/controlpanels/controlpanels.test.ts
@@ -41,4 +41,28 @@ describe('Controlpanels', () => {
 
   it('should update a content rule in the content rules controlpanel', () =>
     testRequest(app, 'controlpanels/patch_content_rule'));
+
+  it('should add an action to a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/post_content_rules_action'));
+
+  it('should add a condition to a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/post_content_rules_condition'));
+
+  it('should update an action to a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/patch_content_rule_action'));
+
+  it('should update a condition to a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/patch_content_rule_condition'));
+
+  it('should delete an action of a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/delete_content_rule_action'));
+
+  it('should delete a condition of a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/delete_content_rule_condition'));
+
+  it('should move an action of a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/move_content_rule_action'));
+
+  it('should move a condition of a content rule in the content rules controlpanel', () =>
+    testRequest(app, 'controlpanels/move_content_rule_condition'));
 });